requirements.txt
aenum==3.1.16
contourpy==1.3.3
cycler==0.12.1
dcps==0.8.0
fonttools==4.61.1
keyboard==0.13.5
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.1
packaging==26.0
pillow==12.1.0
py_kelctl==0.3.2
pyparsing==3.3.2
pyserial==3.5
python-dateutil==2.9.0.post0
pyusb==1.3.1
PyVISA==1.16.1
PyVISA-py==0.8.1
six==1.17.0
typing_extensions==4.15.0
